refactor: address review findings — fix stale docs and reduce test duplication
- Replace all MessengerTransport references with AmqpExchange in docs
- Update CLAUDE.md bridge example to show ContainerInterface sender locator
- Fix curly-brace import violating project convention in CLAUDE.md
- Remove dead amqpConsumeTransport from InboxFlowTestContext
- Extract duplicated createTrackingStack/createPassThroughStack to MiddlewareStackFactory
- Extract duplicated serialiser setup to EventBusFactory::createSerializers()
- Add $logger to DeduplicationMiddleware in test config for env parity
- Apply ECS formatting fixes

Part of #19
