+++ CLAUDE.md
@@ -380,7 +380,7 @@ messenger/
 │   │   │   ├── AmqpRoutingKey.php         # Attribute for custom routing key
 │   │   │   ├── AmqpRoutingStrategyInterface.php
 │   │   │   ├── DefaultAmqpRoutingStrategy.php
-│   │   │   └── MessengerTransport.php     # Attribute for custom exchange
+│   │   │   └── AmqpExchange.php            # Attribute for custom exchange
 │   │   └── MessageName.php                # Attribute for semantic message names
 │   ├── Serializer/                 # Serialization Infrastructure
 │   │   ├── Normalizer/
@@ -634,11 +634,12 @@ services:
     tags:
       - { name: 'messenger.middleware' }
 
-  # Outbox Bridge (middleware — publishes outbox events to AMQP)
-  # Reads existing MessageIdStamp from envelope and publishes via direct SenderInterface
+  # Outbox Bridge (middleware — publishes outbox events to AMQP via sender locator)
+  # Reads existing MessageIdStamp from envelope and publishes via resolved SenderInterface
   Freyr\MessageBroker\Outbox\EventBridge\OutboxToAmqpBridge:
     arguments:
-      $amqpSender: '@messenger.transport.amqp'
+      $senderLocator: !service_locator
+        amqp: '@messenger.transport.amqp'
       $routingStrategy: '@Freyr\MessageBroker\Outbox\Routing\AmqpRoutingStrategyInterface'
       $logger: '@logger'
     tags:
@@ -659,11 +660,12 @@ services:
 ```php
 use Freyr\MessageBroker\Outbox\MessageName;
 use Freyr\MessageBroker\Outbox\EventBridge\OutboxMessage;
-use Freyr\MessageBroker\Outbox\Routing\{MessengerTransport, AmqpRoutingKey};
+use Freyr\MessageBroker\Outbox\Routing\AmqpExchange;
+use Freyr\MessageBroker\Outbox\Routing\AmqpRoutingKey;
 use Freyr\Identity\Id;
 
 #[MessageName('order.placed')]  // REQUIRED: Message name for routing
-#[MessengerTransport('commerce')]     // OPTIONAL: Override default exchange
+#[AmqpExchange('commerce')]     // OPTIONAL: Override default exchange
 #[AmqpRoutingKey('order.test')] // OPTIONAL: Override default routing key
 final readonly class OrderPlaced implements OutboxMessage
 {
@@ -681,7 +683,7 @@ final readonly class OrderPlaced implements OutboxMessage
 3. NO `messageId` property - it's auto-generated by `MessageIdStampMiddleware` at dispatch time as UUID v7
 
 **AMQP Routing (Optional):**
-4. Use `#[MessengerTransport('name')]` to override default exchange (first 2 parts of message name)
+4. Use `#[AmqpExchange('name')]` to override default exchange (first 2 parts of message name)
 5. Use `#[AmqpRoutingKey('key')]` to override default routing key (full message name)
 
 See `docs/amqp-routing.md` for complete routing documentation.
@@ -903,11 +905,14 @@ final readonly class MessageIdStampMiddleware implements MiddlewareInterface
 ```
 
 ```php
-// ✅ Bridge middleware reads existing stamp and publishes to AMQP
+// ✅ Bridge middleware reads existing stamp and publishes to AMQP via sender locator
 final readonly class OutboxToAmqpBridge implements MiddlewareInterface
 {
+    /**
+     * @param ContainerInterface $senderLocator Service locator keyed by transport name (e.g. 'amqp', 'commerce')
+     */
     public function __construct(
-        private SenderInterface $amqpSender,
+        private ContainerInterface $senderLocator,
         private AmqpRoutingStrategyInterface $routingStrategy,
         private LoggerInterface $logger,
         private string $outboxTransportName = 'outbox',
@@ -929,8 +934,10 @@ final readonly class OutboxToAmqpBridge implements MiddlewareInterface
         $messageIdStamp = $envelope->last(MessageIdStamp::class)
             ?? throw new RuntimeException('...');
 
-        // Publish directly to AMQP (no nested bus dispatch)
-        $this->amqpSender->send(new Envelope($event, [
+        // Resolve sender by name and publish to AMQP (no nested bus dispatch)
+        $senderName = $this->routingStrategy->getSenderName($event);
+        $sender = $this->senderLocator->get($senderName);
+        $sender->send(new Envelope($event, [
             $messageIdStamp,
             new AmqpStamp($routingKey, AMQP_NOPARAM, $headers),
         ]));
