+++ README.md
@@ -149,11 +149,11 @@ By default, events are published using convention-based routing:
 You can override this with attributes:
 
 ```php
-use Freyr\MessageBroker\Outbox\Routing\MessengerTransport;
+use Freyr\MessageBroker\Outbox\Routing\AmqpExchange;
 use Freyr\MessageBroker\Outbox\Routing\AmqpRoutingKey;
 
 #[MessageName('order.placed')]
-#[MessengerTransport('commerce')]           // Custom exchange
+#[AmqpExchange('commerce')]                  // Custom exchange
 #[AmqpRoutingKey('commerce.order.placed')]   // Custom routing key
 final readonly class OrderPlaced
 {
@@ -192,7 +192,7 @@ class OrderService
         $this->entityManager->persist($order);
 
         // Dispatch event (saved to outbox in same transaction)
-        // Note: messageId is auto-generated by OutboxToAmqpBridge
+        // Note: messageId is auto-generated by MessageIdStampMiddleware at dispatch
         $this->messageBus->dispatch(new OrderPlaced(
             orderId: $order->getId(),
             customerId: $order->getCustomerId(),
@@ -540,9 +540,10 @@ framework:
     buses:
       messenger.bus.default:
         middleware:
+          - 'Freyr\MessageBroker\Outbox\MessageIdStampMiddleware'  # Stamps OutboxMessage at dispatch
           - doctrine_transaction  # Priority 0 (starts transaction)
-          # DeduplicationMiddleware (priority -10) registered via service tag
-          # Runs after transaction starts, before handlers
+          - 'Freyr\MessageBroker\Outbox\EventBridge\OutboxToAmqpBridge'  # Publishes outbox → AMQP
+          - 'Freyr\MessageBroker\Inbox\DeduplicationMiddleware'  # Inbox deduplication (priority -10)
 
     transports:
       outbox:
@@ -668,11 +669,16 @@ services:
   Freyr\MessageBroker\Outbox\Routing\AmqpRoutingStrategyInterface:
     class: Freyr\MessageBroker\Outbox\Routing\DefaultAmqpRoutingStrategy
 
-  # Outbox Bridge (publishes outbox events to AMQP)
+  # MessageIdStamp Middleware
+  Freyr\MessageBroker\Outbox\MessageIdStampMiddleware:
+    tags:
+      - { name: 'messenger.middleware' }
+
+  # Outbox Bridge (middleware — publishes outbox events to AMQP via sender locator)
   Freyr\MessageBroker\Outbox\EventBridge\OutboxToAmqpBridge:
-    autoconfigure: true
     arguments:
-      $eventBus: '@messenger.default_bus'
+      $senderLocator: !service_locator
+        amqp: '@messenger.transport.amqp'
       $routingStrategy: '@Freyr\MessageBroker\Outbox\Routing\AmqpRoutingStrategyInterface'
       $logger: '@logger'
